move IPAddr related code to dedicated source file ipaddr.py
